Add files via upload
I changed the name of the Purchase price on my Toy Inventory page to Rental Price and put $2.50 as the Rental price for each toy/game, as that is how much it costs to rent out a toy/game. 

Added a drop-down menu for rental prices on the admin page 

Put the logo in the middle on both the admin and website.